Add files via upload
A popular visual of folks who visit the CSI Lab. This document describes the current approach the Lab takes to support a team on it's journey through the lab at a high level. Note that this approach is continuously improving as we learn what works well for teams. 

Some of the resources listed in the CSI lab offering sections may be found in this GitHub Repo.