Upload latest version of Value Canvas Workbook
This is a workbook we offer to people interested in applying the CSI Lab's methods to their work. It is used as a screening tool for potential teams, and is a starting place for a more in depth conversation about what might be possible or needed to build a new service in a digital way.