Replaced previous version
We've updated the name of the lab to reflect our alignment with the BCDevExchange, and have made some improvements to how we communicate the journey of a team at the lab. In particular, we are using similar language to the UK in our process: Align, Define, Deliver (alpha, beta, live).