An illustration of our organization
We have an immediate need to help people understand how our organization is structured, and how the various parts of our programs connect. This is an illustration that communicates what the folks inside the org currently experience and deliver. Of course, we're an Agile team... so as we get feedback on our services and mandate, this could very well change.